Refactor S3 and presigned URL logic to FileUtils
Moved S3 client, presigned URL generation, cache clearing, and temp file cleanup logic into a new FileUtils utility module. Updated Gig and User route handlers to use the shared utilities, reducing code duplication and improving maintainability.
