+++ src/Routes/User.ts
@@ -17,86 +17,10 @@ import {
   updateEmployerProfileSchema
 } from "../Middleware/validator.ts";
 import { uploadDocument, uploadProfilePhoto } from "../Middleware/uploadFile.ts";
-import { file, S3Client, S3File } from "bun";
-import { PresignedUrlCache } from "../Client/RedisClient.ts";
 import { Role } from "../Types/types.ts";
-
-const s3Client = new S3Client({
-  region: "auto",
-  accessKeyId: process.env.R2ACCESSKEYID,
-  secretAccessKey: process.env.R2SECRETACCESSKEY,
-  endpoint: process.env.R2ENDPOINT,
-  bucket: "backend-files",
-  retry: 1,
-});
-
-const R2_BUCKET_NAME = "backend-files";
-
-async function generatePresignedUrl(filePath: string, expiresIn = 3600): Promise<string | null> {
-  if (!filePath) {
-    console.warn("generatePresignedUrl called with empty filePath");
-    return null;
-  }
-
-  const cacheKey = `presigned:${R2_BUCKET_NAME}:${filePath}`;
-
-  try {
-    const cachedUrl = await PresignedUrlCache.get(cacheKey);
-    if (cachedUrl) {
-      // PresignedUrlCache.get already checks for near expiry
-      console.log(`Returning cached presigned URL for ${filePath}`);
-      return cachedUrl;
-    }
-  } catch (error) {
-    console.error(`Error fetching from PresignedUrlCache for ${cacheKey}:`, error);
-    // Proceed to generate a new URL if cache fetch fails
-  }
-
-  try {
-    // documentation and updated if necessary.
-    // The exact structure of the returned value (string or object with a URL property)
-    const signedRequestResult = await (s3Client as any).presign(filePath, {
-      expires: expiresIn,
-    });
-
-    // Adjust if 'signedRequestResult' is an object (e.g., signedRequestResult.url)
-    const finalUrl = typeof signedRequestResult === 'object' && signedRequestResult.url ? signedRequestResult.url : signedRequestResult;
-
-    if (!finalUrl) {
-      console.error(`Failed to generate presigned URL for ${filePath}. Method might be incorrect or returned null/undefined.`);
-      return null;
-    }
-
-    await PresignedUrlCache.set(cacheKey, finalUrl, expiresIn);
-    console.log(`Successfully generated and cached presigned URL for ${filePath}`);
-    return finalUrl;
-
-  } catch (error) {
-    console.error(`Error generating presigned URL for ${filePath}:`, error);
-    return null;
-  }
-}
+import { s3Client, generatePresignedUrl, cleanupTempFiles } from "../Utils/FileUtils.ts";
 
 const router = new Router();
-
-// 清理臨時文件
-const cleanupTempFiles = async (uploadedFiles: any[]) => {
-  if (uploadedFiles.length === 0) return;
-
-  Promise.all(
-    uploadedFiles.map(async (file) => {
-      try {
-        const bunFile = Bun.file(file.path);
-        if (await bunFile.exists()) {
-          await bunFile.delete();
-          console.log(`成功刪除臨時文件: ${file.filename}`);
-        }
-      } catch (cleanupError) {
-        console.error(`清理臨時文件時出錯 ${file.filename}:`, cleanupError);
-      }
-    })
-  ).catch(err => console.error('批次清理檔案時出錯:', err));
-};
 
 router.post(
   "/register/worker",
